Cont. see for details... - Hunk implementation wasn't platform specific; moved into hunk.cpp - Made a start looking at mouse handling - Continued gutting sdl2_sys.cpp as it's migrated into app.cpp - We no longer attempt to download data on localhost

+++ game/q_shared.h
@@ -273,7 +273,7 @@ void Sys_Mkdir(char *path);
 void *Hunk_Begin( size_t maxsize);
 void *Hunk_Alloc( size_t size);
 void Hunk_Free(void *buf);
-int Hunk_End();
+size_t Hunk_End();
 
 // directory searching
 #define SFF_ARCH 0x01
